Refactor hooks
* Make a new directory for hooks.
* Make hookset a real well-defined class interface.
* Make SlackHooks an implementation of that interface.
* Get rid of globalHooks, import GlobalHooks
